requirements.txt
moviepy
mayavi
numpy
matplotlib
torch
hydra-core
